modifying the git workflow section of the README
Since this is the first time we have worked with branches, I am writing out all of the commands we need to branch and merge.  I have completed the basic 'how to merge a branch if everyone is working on sepearate files' section. I have not dealt with basic merge conflicts.